Creating ways to calculate GameScore
GameScore is a stat that is used a lot in the NBA, and since it isn't really a stat that is used by default there isn't a clear way to calculate this property. Which gives a lot of room to create new ways of calculating this. Which is what this NoteBook will serve.